Move the citation information out of the about tab into it's own spot.
